+++ weather-app/src/hooks/useDetails.ts
