+++ weather-app/src/main.tsx
@@ -2,16 +2,16 @@ import { createRoot } from "react-dom/client";
 
 import { BrowserRouter, Navigate, Route, Routes } from "react-router";
 import { Navbar } from "./components/Navbar.tsx";
-import { Home } from "./pages/home/Home.tsx";
-import { Forecasts } from "./pages/forecasts/Forecasts.tsx";
-import { Favorites } from "./pages/favorites/Favorites.tsx";
+import { Home } from "./pages/Home.tsx";
+import { Forecasts } from "./pages/Forecasts.tsx";
+import { Favorites } from "./pages/Favorites.tsx";
 import { NotFound } from "./pages/NotFound.tsx";
 import { Routes as RoutesEnum } from "./routes.ts";
-import { ThemeProvider } from "./contexts/theme/ThemeContext.tsx";
+import { ThemeProvider } from "./contexts/ThemeContext.tsx";
 import { ThemeSwitcher } from "./components/buttons/ThemeSwitcher.tsx";
 import { Provider } from "react-redux";
 import { store } from "./store.ts";
-import { Details } from "./pages/forecasts/details/Details.tsx";
+import { Details } from "./pages/Details.tsx";
 import ErrorBoundary from "./components/ErrorBoundary.tsx";
 
 createRoot(document.getElementById("root")!).render(
