+++ weather-app/src/slices/favoritesSlice.ts
