+++ weather-app/src/slices/searchSlice.ts
